Start onboarding only if logged in
